src/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.1
kombu==4.2.0
msgpack-python==0.5.6
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
django-enumfields==2.0.0
drf-yasg==1.17.0
gunicorn==19.9.0
psycopg2==2.7.5
requests==2.23.0
